+++ vscode/microsoft-kiota/src/constants.ts
@@ -15,3 +15,4 @@ export const APIMANIFEST = "apimanifest";
 export const REMIND_ME_LATER_FLAG = 'remindMeLater';
 export const API_MANIFEST_FILE = "apimanifest.json";
 export const MANIFEST_KIOTA_VERSION_KEY = "x-ms-kiota-version";
+export const SHOW_MESSAGE_AFTER_API_LOAD = "show-message-after-api-load";